# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Valley MAS 创作者内容平台 API 文档\n支持创作者注册、口令管理、资源上传下载等功能",
    "title": "Valley MAS API",
    "contact": {
      "name": "API Support",
      "email": "support@valley-mas.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "paths": {
...
    "/admin/records/downloads": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "管理员查看所有下载记录",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 记录管理"],
        "summary": "获取下载记录",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 20,
            "description": "每页数量",
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "下载记录列表",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/admin/records/uploads": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "管理员查看所有上传记录",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 记录管理"],
        "summary": "获取上传记录",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 20,
            "description": "每页数量",
            "name": "pageSize",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "上传记录列表",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/admin/resources": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "管理员查看所有资源列表",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 资源管理"],
        "summary": "获取资源列表",
        "parameters": [
          {
            "type": "integer",
            "default": 1,
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 20,
            "description": "每页数量",
            "name": "pageSize",
            "in": "query"
          },
          {
            "enum": ["avatar", "wallpaper"],
            "type": "string",
            "description": "资源类型",
            "name": "type",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "资源列表",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/admin/resources/upload": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "管理员上传头像或壁纸资源",
        "consumes": ["multipart/form-data"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 资源管理"],
        "summary": "上传资源",
        "parameters": [
          {
            "type": "file",
            "description": "资源文件",
            "name": "file",
            "in": "formData",
            "required": true
          },
          {
            "enum": ["avatar", "wallpaper"],
            "type": "string",
            "description": "资源类型",
            "name": "type",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "上传成功",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "/admin/resources/{id}": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "管理员删除资源",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["管理后台 - 资源管理"],
        "summary": "删除资源",
        "parameters": [
          {
            "type": "string",
            "description": "资源ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "401": {
            "description": "未登录",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "403": {
            "description": "无权限",
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT 认证，格式：Bearer {token}",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}